Added Create Date helper function
