+++ src/components/navbar/Navbar.jsx
@@ -1,4 +1,4 @@
-import { LOGO } from "src/utils/constants/constants";
+import { NAV_LOGO } from "src/utils/constants/constants";
 import Search from "src/components/search/Search";
 import "src/components/navbar/Navbar.component.scss";
 
@@ -6,7 +6,7 @@ const Navbar = () => {
   return (
     <div className="navbar">
       <div className="navbar-logo">
-        <img src={LOGO} alt="Logo" />
+        <img src={NAV_LOGO} alt="Logo" />
         <h1>Todos</h1>
       </div>
       <div className="search-bar">
