+++ src/utils/constants/constants.jsx
@@ -1,15 +1,17 @@
 import mark from "src/assets/images/mark.png";
 import pen from "src/assets/images/pen.png";
 import bin from "src/assets/images/bin.png";
-import logo from "src/assets/images/nav-logo.png";
+import logo from "src/assets/images/logo.png";
+import navLogo from "src/assets/images/nav-logo.png";
 import search from "src/assets/images/search.png";
 import addicon from "src/assets/images/add.png";
 
-export const [MARK_ICON, PEN_ICON, BIN_ICON, LOGO, SEARCH_ICON, ADD_ICON] = [
-  mark,
-  pen,
-  bin,
-  logo,
-  search,
-  addicon,
-];
+export const [
+  MARK_ICON,
+  PEN_ICON,
+  BIN_ICON,
+  LOGO,
+  NAV_LOGO,
+  SEARCH_ICON,
+  ADD_ICON,
+] = [mark, pen, bin, logo, navLogo, search, addicon];
